cleaned up Hsv to Hex conversion
